requirements.txt
python-telegram-bot
python-dotenv
anthropic
psycopg2-binary
sqlalchemy
